requirements.txt
Flask==2.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
python-decouple
python-dotenv==0.20.0
pymongo==4.1.1
